Add store to main electron script
Store class provides methods for accessing and overwriting configuration data. Normally, this class is separated into a dedicated module file, but recent attempts reveal some conflict with the fs module. It may be the case that the fs module can only be used in the main script for security reasons. Ideally, this will be changed in future development.